stable widgets yayaaay
git-svn-id: https://paroxysm.googlecode.com/svn/trunk@114 1a4693bc-ef48-11de-b206-85b67c77eac7

+++ source/Button.h
@@ -54,8 +54,8 @@ class Button
 
         Point2D<float> mLocation;
         Point2D<float> mSize;
-        Point2D<float> mPixelUL;
-        Point2D<float> mPixelLR;
+        Point2D<int> mPixelUL;
+        Point2D<int> mPixelLR;
 };
 
 inline bool Button::isOver(int inX, int inY)
